Add stuff related to ModBus TCP/IP using pymodbus
I tried the free chatGPT 3.5 and COPILOT's chatGPT 4.0, but what worked best was the Example on the PyModBus website... I mixed a little thing or two and made it compilable for PyInstaller. I leave it here in case anyone needs it or I forget it myself. Good version for ModBus TCP/IP communication tests, being able to choose IP and Coil and HR registers for test readings.