Add code to insert data into PostgreSQL
This script must be called via the command line, arguments 1 and 2 will be the values to be inserted into the database. You need to define in code: database name, user, password, host, port (if different from default), table name, column names and add arguments if there are more than 2 values.